+++ maeil-mail/src/main/java/maeilmail/admin/AdminNoticeService.java
@@ -2,6 +2,7 @@ package maeilmail.admin;
 
 import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
+import maeilmail.bulksend.sender.AdminNoticeSender;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
